#39 Refactored TestInvoker to handle the case in which the test method is async. Added tests for notifications when exceptions are throws from async methods.
